[fix] : 코드 리펙토링 밑 패키지 분리
