+++ src/main/java/SpotifyPlaylist/Controller/SpotifyController.java
@@ -1,28 +1,32 @@
 package SpotifyPlaylist.Controller;
 
-import SpotifyPlaylist.Service.SpotifyService;
-import org.apache.hc.core5.http.ParseException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import se.michaelthelin.spotify.exceptions.SpotifyWebApiException;
-import se.michaelthelin.spotify.model_objects.specification.Artist;
-import se.michaelthelin.spotify.model_objects.specification.Track;
-
+import SpotifyPlaylist.DTO.SearchResponseDto;
+import SpotifyPlaylist.SpotifyConfig;
 import java.io.IOException;
 import java.util.List;
-
+import org.apache.hc.core5.http.ParseException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RestController;
+import se.michaelthelin.spotify.SpotifyApi;
+import se.michaelthelin.spotify.exceptions.SpotifyWebApiException;
+import SpotifyPlaylist.Service.SpotifyService;
 @RestController
 public class SpotifyController {
+    SpotifyApi spotifyApi = new SpotifyApi.Builder()
+            .setAccessToken(SpotifyConfig.accessToken())
+            //accessToken()은 SpotifyConfig 에서 발급함
+            .build();
+
     @Autowired
-    private SpotifyService spotifyService;
+    SpotifyService spotifyService =new SpotifyService();
 
-    @GetMapping("/tracks/{artistName}")
-    public List<Track> searchTracksByArtist(@PathVariable String artistName) throws IOException, ParseException, SpotifyWebApiException {
-        return spotifyService.searchTracksByArtist(artistName);
+    @GetMapping("/search/{trackname}")
+    public List<SearchResponseDto> searchTracksByArtist(@PathVariable String trackname) throws IOException, ParseException, SpotifyWebApiException {
+
+
+        return spotifyService.search(trackname);
     }
-
-
 
 }
