--- src/main/java/SpotifyPlaylist/spotifyconfig/SpotifyService1.java
@@ -1,56 +1,0 @@
-package SpotifyPlaylist.spotifyconfig;
-
-import SpotifyPlaylist.spotifyconfig.DTO.SearchResponseDto;
-import java.io.IOException;
-import java.text.ParseException;
-import java.util.ArrayList;
-import java.util.List;
-import se.michaelthelin.spotify.SpotifyApi;
-import se.michaelthelin.spotify.exceptions.SpotifyWebApiException;
-import se.michaelthelin.spotify.model_objects.specification.AlbumSimplified;
-import se.michaelthelin.spotify.model_objects.specification.ArtistSimplified;
-import se.michaelthelin.spotify.model_objects.specification.Image;
-import se.michaelthelin.spotify.model_objects.specification.Paging;
-import se.michaelthelin.spotify.model_objects.specification.Track;
-import se.michaelthelin.spotify.requests.data.search.simplified.SearchTracksRequest;
-
-public class SpotifyService1 {
-    SpotifyApi spotifyApi = new SpotifyApi.Builder()
-            .setAccessToken(SpotifyConfig.accessToken())
-            //accessToken()은 SpotifyConfig 에서 발급함
-            .build();
-
-    public List<SearchResponseDto> search(String keyword) {
-        List <SearchResponseDto> searchResponseDtoList = new ArrayList<>();
-
-        try {
-            SearchTracksRequest searchTrackRequest = spotifyApi.searchTracks(keyword)
-                    .limit(10)
-                    .build();
-
-            Paging<Track> searchResult = searchTrackRequest.execute();
-            Track[] tracks = searchResult.getItems();
-
-            for (Track track : tracks) {
-                String title = track.getName();
-
-                AlbumSimplified album = track.getAlbum();
-                ArtistSimplified[] artists = album.getArtists();
-                String artistName = artists[0].getName();
-
-
-                Image[] images = album.getImages();
-                String imageUrl = (images.length > 0) ? images[0].getUrl() : "NO_IMAGE";
-
-                String albumName = album.getName();
-
-                searchResponseDtoList.add(DtoMapper.toSearchDto(artistName, title, albumName, imageUrl));
-            }
-
-        } catch (IOException | SpotifyWebApiException | org.apache.hc.core5.http.ParseException e) {
-            System.out.println("Error: " + e.getMessage());
-        }
-        return searchResponseDtoList;
-    }
-
-}
